Update WA.cpp
Add in new functions which are commented out as their current versions prevent compiling due to call issues. Need to write simpler versions of the functions such as the task 3 function, route 3 finder, and adjChck functions which are used for the task 3 prompt, route finding and building and the adjacency check for two given int values.