Update adjacencyList.cpp
updated to utilize BFS search for task 3. BFS search function does not work currently as there is an issue with one of the while loops causing a stall in the code but the code compiles without errors. BFS is being used instead of DFS as BFS is faster and is better for finding the shortest path in a given search.